Added audio information, and expanded the reset information.
I tested both Google Stadia and the DualShock 4, and they both support
audio input and output (i.e. TRRS, headset). However, when trying, I
noticed some audio delay.

My system is Manjaro Linux with Pipewire. I tested the controllers
connected directly to the front USB ports of the computer. I used
qpwgraph to rewire the microphones to the headphones, and I noticed it
was hard to speak properly due to the delay.

Then I used https://friture.org/ and qpwgraph again to rewire two
microphones as left and right channels. Friture was then able to
calculate the delay between those channels.

I compared the gamepad microphone inputs to the on-board analog audio
input of my computer. I also compared it to the USB microphone I use.

It is possible some of the delay is due to pipewire, or due to the
non-realtime Linux kernel. My test wasn't perfect, and I didn't spend
much time replicating the test for precise results. Maybe later.